+++ app/Models/Site.php
@@ -131,6 +131,11 @@ class Site extends Model
         return $this->hasMany(SiteMalwareScan::class)->latest();
     }
 
+    public function applications(): HasMany
+    {
+        return $this->hasMany(SiteApplication::class)->orderBy('type');
+    }
+
     public function gitRepository(): HasOne
     {
         return $this->hasOne(SiteGitRepository::class);